% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_layout.R
\name{delete_layout}
\alias{delete_layout}
\title{Delete an existing project}
\usage{
delete_layout(layout_name)
}
\arguments{
\item{layout_name}{A string denoting the name of the pubSEM project to
delete.}
}
\value{
nothing
}
\description{
Delete an existing project
}
