% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_project.R
\name{delete_project}
\alias{delete_project}
\title{Delete an existing project}
\usage{
delete_project(project_name)
}
\arguments{
\item{project_name}{A string denoting the name of the pubSEM project to
delete.}
}
\value{
nothing
}
\description{
Delete an existing project
}
