% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_diagram.R
\name{export_diagram}
\alias{export_diagram}
\title{Export a pubSEM layout to PDF}
\usage{
export_diagram(layout_name, filename, directory = getwd())
}
\arguments{
\item{layout_name}{a string denoting the pubSEM layout to export}

\item{filename}{a string specifying the name of the exported PDF}

\item{directory}{a string specifying the directory in which to save the
exported PDF. Defaults to the current working directory.}
}
\value{
nothing
}
\description{
Export a pubSEM layout to PDF
}
