% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layout_directory.R
\name{get_layout_directory}
\alias{get_layout_directory}
\title{Get the directory holding the layout files}
\usage{
get_layout_directory()
}
\value{
a file path to the directory containing pubSEM layout files
}
\description{
Get the directory holding the layout files
}
