% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubSEM-package.R
\docType{package}
\name{pubSEM-package}
\alias{pubSEM}
\alias{pubSEM-package}
\title{pubSEM: Create Publication-Ready Path Diagrams from 'lavaan' Models}
\description{
Easily create aesthetically pleasing and reproducable path diagrams from fitted 'lavaan' structural equation models using an included GUI.
}
\author{
\strong{Maintainer}: Dylan Glenn \email{dylanwglenn@gmail.com}

}
\keyword{internal}
