% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_gui.R
\name{sem_gui}
\alias{sem_gui}
\title{Open the pubSEM GUI editor}
\usage{
sem_gui(fit, project_name, standardized = FALSE)
}
\arguments{
\item{fit}{A lavaan object}

\item{project_name}{A string denoting the name of the pubSEM project to which
a layout should be stored. A project file stores persistent data for
reproducible diagrams in the current working directoloadry — project names
should be unique EVEN ACROSS R PROJECTS!}

\item{standardized}{A bool value}
}
\value{
nothing
}
\description{
Open the pubSEM GUI editor
}
