% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_gui.R
\name{sem_gui}
\alias{sem_gui}
\title{Open the pubSEM GUI editor}
\usage{
sem_gui(fit, layout_name, standardized = FALSE)
}
\arguments{
\item{fit}{A lavaan object}

\item{layout_name}{A string denoting the name of the pubSEM layout to which
a layout should be stored. A layout file stores persistent data for
reproducible diagrams in the current working directoloadry — layout names
should be unique EVEN ACROSS R projects!}

\item{standardized}{A bool value}
}
\value{
nothing
}
\description{
Open the pubSEM GUI editor
}
